Modified files to suit changes

+++ src/main.rs
@@ -9,36 +9,38 @@ use remove_from_array::remove_index;
 use reverse_array::reverse_array;
 
 fn main() {
-    
-    
     println!("\nPlease enter the number element of the array [1 <= number of elements <= 50]");
     let mut num_of_elem = get_nums();
     let mut arr = get_array(num_of_elem);
-loop {
-    println!("\nPlease Enter\n1=>To reverse the array\n2=>To remove an element from the array\n3=>To add an element into the array\n4=>To Exit");
-
-    let opt = get_nums();
-
-    match opt {
-        1 => {
-            arr = reverse_array(num_of_elem, arr);
-        }
-        2 => {
-            println!("\nPlease enter the position at which you would like to remove an element from the initail array [ 0 < index < number of elements ]");
-            let index = get_nums();
-            (arr, num_of_elem) = remove_index(index-1, arr, num_of_elem);
-        }
-        3 => {
-            println!("\nPlease enter the position at which you would like to insert an element [ 0 < index < number of elements ]");
-            let index = get_nums();
-            println!("\nEnter the element you would like to insert");
-            let element = get_nums();
-           (arr, num_of_elem) = add_element(index-1, element, num_of_elem, arr);
-        },
-        4 => {break;},
-        _ => println!("Please enter one of the above options"),
+    if num_of_elem < 1 {
+        return println!("Number of element cannot be less than 1");
     }
-    println!("\nEnter 4 to quite the program");
-    if opt == 4 { break;}
-}
+    loop {
+        println!("\nPlease Enter\n1=>To reverse the array\n2=>To remove an element from the array\n3=>To add an element into the array\n4=>To Exit");
+
+        let opt = get_nums();
+
+        match opt {
+            1 => {
+                arr = reverse_array(num_of_elem, arr);
+            }
+            2 => {
+                println!("\nEnter the position  to remove  element from array [ 0 < index < number of elements ]");
+                let index = get_nums();
+                (arr, num_of_elem) = remove_index(index - 1, arr, num_of_elem);
+            }
+            3 => {
+                println!("\nEnter the position to insert new element [ 0 < index < number of elements ]");
+                let index = get_nums();
+                println!("\nEnter the element you would like to insert");
+                let element = get_nums();
+                (arr, num_of_elem) = add_element(index - 1, element, num_of_elem, arr);
+            }
+            4 => {
+                break;
+            }
+            _ => println!("Sorry, enter one of the above options to proceed!"),
+        }
+        println!();
+    }
 }
